xsearch/requirements.txt
# 智能分析系统依赖包
google-generativeai>=0.3.0
langextract>=0.1.0
pyyaml>=6.0
llama-index>=0.10.0
llama-index-vector-stores-faiss>=0.1.0
llama-index-embeddings-dashscope>=0.1.0
llama-index-llms-openai-like>=0.1.0
faiss-cpu>=1.7.0
